# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user": {
            "post": {
                "description": "Creates a user with the given input and returns an array of user objects, containing the newly-created user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "admin-only operation to create a user",
                "parameters": [
                    {
                        "description": "The input user data (only `first_name`, `last_name` and `email` are required)",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an user and returns a response to indicate success or failure",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "admin-only operation to delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a user with the given input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "updates a user",
                "parameters": [
                    {
                        "description": "The input user update data (only `id` is required, but is not useful without setting other fields to update)",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "description": "Gets user(s) by the ID(s) in the request query string, `?ids=`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "gets user(s)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "user search by id",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.V1_API_RESPONSE_USERS"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Entree": {
            "type": "object",
            "required": [
                "option_name"
            ],
            "properties": {
                "created_at": {
                    "description": "The time the record was created at\n\nWe override Gorm's CreatedAt field so we can set the gorm:\"\u003c-:create\" directive,\nwhich prevents this field from being altered once the record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "option_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.HorsDoeuvres": {
            "type": "object",
            "required": [
                "option_name"
            ],
            "properties": {
                "created_at": {
                    "description": "The time the record was created at\n\nWe override Gorm's CreatedAt field so we can set the gorm:\"\u003c-:create\" directive,\nwhich prevents this field from being altered once the record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "option_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name"
            ],
            "properties": {
                "created_at": {
                    "description": "The time the record was created at\n\nWe override Gorm's CreatedAt field so we can set the gorm:\"\u003c-:create\" directive,\nwhich prevents this field from being altered once the record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "description": "The user's email (must be unique); this field is an index.",
                    "type": "string"
                },
                "entreeSelection": {
                    "$ref": "#/definitions/models.Entree"
                },
                "entree_selection_id": {
                    "description": "The ID of the entree the user has selected; is null until the user makes a selection.",
                    "type": "string"
                },
                "first_name": {
                    "description": "The user's first name.",
                    "type": "string"
                },
                "horsDoeuvresSelection": {
                    "$ref": "#/definitions/models.HorsDoeuvres"
                },
                "hors_doeuvres_selection_id": {
                    "description": "The ID of the hors doeuvres the user has selected; is null until the user makes a selection.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_going": {
                    "description": "Whether or not the user is attending.",
                    "type": "boolean"
                },
                "last_name": {
                    "description": "The user's last name.",
                    "type": "string"
                },
                "password": {
                    "description": "Either a hash of the user's password (when stored in the DB), or a plain-text representation of the password (plain-text version is never stored)",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "The user's auth refresh token.",
                    "type": "string"
                },
                "role": {
                    "description": "The user's role, which can be \"GUEST\", \"INVITEE\" or \"ADMIN\". Defaults to \"GUEST\".",
                    "type": "string"
                },
                "token": {
                    "description": "The user's auth token.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "types.UserData": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
...
        "types.V1_API_RESPONSE_USERS": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.UserData"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}